# openapi.yaml
openapi: 3.1.0
info:
  title: Ticket API
  version: 0.1.891
  description: |
    RESTful API for the Ticket project and issue management system.
    All endpoints are under `/api/`. Authentication uses Bearer tokens
    or session cookies (`__Host-session` on secure requests; legacy
    `ticket_token` remains accepted).
    Agent endpoints use HTTP Basic Auth (agent UUID + password).
  contact:
    name: Simon Gauld
    url: https://github.com/simonski/ticket
  license:
    name: MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: Session token from /api/login
    CookieAuth:
      type: apiKey
      in: cookie
      name: __Host-session
    BasicAuth:
      type: http
      scheme: basic
      description: Agent UUID as username, agent password as password
  schemas:
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
      example:
        error: "ticket not found"
    User:
      type: object
      example:
        user_id: "a1b2c3d4-e5f6-7890-abcd-ef1234567890"
        username: "alice"
        role: "user"
        display_name: "Alice Smith"
        enabled: true
        user_type: "user"
      properties:
        user_id:
          type: string
          format: uuid
        username:
          type: string
        role:
          type: string
          enum: [admin, user]
        display_name:
          type: string
        enabled:
          type: boolean
        user_type:
          type: string
          enum: [user, agent]
        email:
          type: string
        status:
          type: string
        description:
          type: string
        last_seen:
          type: string
          format: date-time
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    Project:
      type: object
      properties:
        project_id:
          type: integer
          format: int64
        prefix:
          type: string
          description: 1-5 uppercase ASCII letters
          pattern: '^[A-Z]{1,5}$'
        title:
          type: string
        description:
          type: string
        acceptance_criteria:
          type: string
        git_repository:
          type: string
        notes:
          type: string
        status:
          type: string
          enum: [open, closed]
        visibility:
          type: string
          enum: [public, private]
        workflow_id:
          type: integer
          format: int64
          nullable: true
        ticket_sequence:
          type: integer
        created_by:
          type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    StatusResponse:
      type: object
      properties:
        status:
          type: string
        authenticated:
          type: boolean
        registration_enabled:
          type: boolean
        chat_enabled:
          type: boolean
        chat_max_connections:
          type: integer
        chat_max_duration_minutes:
          type: integer
        chat_running_processes:
          type: integer
        server_version:
          type: string
        user:
          $ref: '#/components/schemas/User'
...
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required or token invalid
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    NotFound:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    InternalServerError:
      description: Unexpected server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
  examples:
...
    ProjectExample:
      summary: A typical project
      value:
        project_id: 1
        prefix: "CUS"
        title: "Customer Portal"
        description: "Customer-facing web portal project"
        status: "open"
        visibility: "private"
        created_at: "2025-01-15T10:00:00Z"
        updated_at: "2025-06-20T14:30:00Z"
...
    ProjectListExample:
      summary: Project list response
      value:
        - project_id: 1
          prefix: "CUS"
          title: "Customer Portal"
          description: "Customer-facing web portal project"
          status: "open"
          visibility: "private"
        - project_id: 2
          prefix: "OPS"
          title: "Operations Console"
          description: "Internal operational tooling"
          status: "open"
          visibility: "private"
    CreateProjectRequestExample:
      summary: Create project request
      value:
        prefix: "CUS"
        title: "Customer Portal"
        description: "Customer-facing web portal project"
        visibility: "private"
        workflow_id: 1
...
paths:
...
  /api/status:
    get:
      tags: [Health]
      summary: Server status and feature flags
      operationId: getStatus
      description: Returns server status. Includes user info if authenticated.
      responses:
        '200':
          description: Status information
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StatusResponse'

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
  # ── Authentication ──────────────────────────────────────────────────
...
  /api/projects:
    get:
      tags: [Projects]
      summary: List projects visible to the authenticated user
      operationId: listProjects
      security:
        - BearerAuth: []
      responses:
        '200':
          description: Project list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Project'
              examples:
                default:
                  $ref: '#/components/examples/ProjectListExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [prefix, title]
              properties:
                prefix:
                  type: string
                  description: 1-5 uppercase ASCII letters
                  pattern: '^[A-Z]{1,5}$'
                title:
                  type: string
                description:
                  type: string
                acceptance_criteria:
                  type: string
                git_repository:
                  type: string
                notes:
                  type: string
                visibility:
                  type: string
                  enum: [public, private]
                workflow_id:
                  type: integer
                  format: int64
            examples:
              default:
                $ref: '#/components/examples/CreateProjectRequestExample'
      responses:
        '201':
          description: Project created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                created:
                  $ref: '#/components/examples/ProjectExample'

        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/projects/{project_id}:
    get:
      tags: [Projects]
      summary: Get project details
      operationId: getProject
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Project details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                default:
                  $ref: '#/components/examples/ProjectExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
    put:
      tags: [Projects]
      summary: Update a project
      operationId: updateProject
      security:
        - BearerAuth: []
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                acceptance_criteria:
                  type: string
                git_repository:
                  type: string
                notes:
                  type: string
                status:
                  type: string
                  enum: [open, closed]
                visibility:
                  type: string
                  enum: [public, private]
                workflow_id:
                  type: integer
                  format: int64
            examples:
              default:
                value:
                  title: "Customer Portal"
                  description: "Customer-facing web portal project"
                  visibility: "private"
                  workflow_id: 1
      responses:
        '200':
          description: Project updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Project'
              examples:
                updated:
                  $ref: '#/components/examples/ProjectExample'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
...